1. feat!: 升级Spring Boot 2.4.10 → 4.1.0，JDK 1.8 → 17，最低支持JDK 17 2. refactor: javax→jakarta命名空间迁移，适配Boot 4废弃/删除API（Base64Utils/UTF8 MediaType/sql init/@WebFilter/selectCount返回Long） 3. chore: 移除JDK 8启动脚本，裁剪bin脚本中JDK 8兼容分支 4. docs: README与开发/打包/部署文档更新JDK 17、Spring Boot 4.1.0要求

+++ src/main/java/com/xuxd/kafka/console/dao/init/DataInit.java
@@ -61,12 +61,12 @@ public class DataInit implements SmartInitializingSingleton {
         }
         try {
             Connection connection = dataSource.getConnection();
-            Integer userCount = userMapper.selectCount(null);
+            Long userCount = userMapper.selectCount(null);
             if (userCount == null || userCount == 0) {
                 initData(connection, SqlParse.USER_TABLE);
             }
 
-            Integer roleCount = roleMapper.selectCount(null);
+            Long roleCount = roleMapper.selectCount(null);
             if (roleCount == null || roleCount == 0) {
                 initData(connection, SqlParse.ROLE_TABLE);
             }
@@ -74,7 +74,7 @@ public class DataInit implements SmartInitializingSingleton {
             if (authConfig.isReloadPermission()) {
                 permissionMapper.delete(null);
             }
-            Integer permCount = permissionMapper.selectCount(null);
+            Long permCount = permissionMapper.selectCount(null);
             if (permCount == null || permCount == 0) {
                 initData(connection, SqlParse.PERM_TABLE);
             }
